+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- `ProbabilityRatioError` gains the following traits
+  - `Display`
+  - `Error`
+
 ## [0.7.0] (2022-09-05)
 
 ### Added